websites-chatbot/requirements.txt
requests
beautifulsoup4
openai
python-dotenv
faiss-cpu   
langchain-openai
langchain-community
tiktoken